Move StorageValue to core.
